+++ CHANGELOG.rst
@@ -6,6 +6,12 @@
 Change Log
 ::::::::::
 
+0.28.0 (u)
+======
+
+* Folio introduces named page styles for front and rear matter.
+* Repin to turberfield-dialogue 0.47.0.
+
 0.27.0
 ======